% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check_coeff.R
\name{Coeff_ErrorCheck}
\alias{Coeff_ErrorCheck}
\title{Error Check for Coefficient in Regression}
\usage{
Coeff_ErrorCheck(data_YZA_s, data_ZA_ns, prop_check = FALSE)
}
\arguments{
\item{data_YZA_s}{list for selected sample (Microdata or Summary statistics). See details below for more details.}

\item{data_ZA_ns}{list for non-selected sample (Microdata or Summary statistics). See details below for more details.}

\item{prop_check}{if \code{TRUE}, check if Y is binary for binary outcome. Default is \code{FALSE}.}
}
\value{
A list of data_YZA_s, sumry_YZA_s and sumry_ZA_ns with
\itemize{
\item \code{data_YZA_s=data_YZA_s} : the same as the input data_YZA_s
\item \code{sumry_YZA_s=list(mean_YZA, var_YZA, n_YZA)} : summary statistics for selected sample
\item \code{sumry_ZA_ns=list(mean_ZA, var_ZA, n_ZA)} : summary statistics for non-selected sample
}
}
\description{
\code{Coeff_ErrorCheck} examinate if the input data for calculating coefficient in regression under proxy pattern-mixture model is correct.
}
\details{
\describe{
\item{\code{data_YZA_s=list(Y,Z,A)}(Microdata)}{
\itemize{
\item \code{Y} : \code{n*1} matrix of outcome variable \code{Y}
\item \code{Z} : \code{n*nZvars} matrix of predictor variables \code{Z}
\item \code{A} : \code{n*nAvars}matrix of auxiliary variables \code{A}
}
}
\item{\code{data_YZA_s=list(mean_YZA, var_YZA, n_YZA)}(Summary Statistics)}{
\itemize{
\item \code{mean_YZA} : \code{(1+nZvars+nAvars)*1} vector of mean (in order) of outcome variable \code{Y}, predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{var_YZA} : \code{(1+nZvars+nAvars)*(1+nZvars+nAvars)} matrix of variance (in order) of outcome variable \code{Y}, predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{n_YZA} : number of selected sample
}
}

\item{\code{data_ZA_ns=list(Z,A)}(Microdata)}{
\itemize{
\item \code{Z} : \code{(N-n)*nZvars} matrix of predictor variables \code{Z}
\item \code{A} : \code{(N-n)*nAvars}matrix of auxiliary variables \code{A}
}
}
\item{\code{data_ZA_ns=list(mean_ZA, var_ZA, n_ZA)}(Summary Statistics)}{
\itemize{
\item \code{mean_ZA} : \code{(nZvars+nAvars)*1} vector of mean (in order) of predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{var_ZA} : \code{(nZvars+nAvars)*(nZvars+nAvars)} matrix of variance (in order) of predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{n_ZA} : number of non-selected sample
}
}

}
}
