% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check.R
\name{ErrorCheck}
\alias{ErrorCheck}
\title{Check if error in the input data for calculating outcome mean estimates under proxy pattern-mixture model}
\usage{
ErrorCheck(data_YZ_s, data_Z_ns, prop_check = FALSE, mi_check = FALSE)
}
\arguments{
\item{data_YZ_s}{list for selected sample (Microdata or Summary statistics). See details below for more details.}

\item{data_Z_ns}{list for non-selected sample (Microdata or Summary statistics). See details below for more details.}

\item{prop_check}{if \code{TRUE}, check if Y is binary for binary outcome. Default is \code{FALSE}.}

\item{mi_check}{if \code{TRUE}, check if data_YZ_s and data_Z_ns are both microdata for multiple imputation. Default is \code{FALSE}.}
}
\value{
A list of data_YZ_s, data_Z_ns, sumry_YZ_s and sumry_Z_ns with
\itemize{
\item \code{data_YZ_s=data_YZ_s} : the same as the input data_YZ_s
\item \code{data_Z_ns=data_Z_ns} : the same as the input data_Z_ns
\item \code{sumry_YZ_s=list(mean_YZ, var_YZ, n_YZ)} : summary statistics for selected sample
\item \code{sumry_Z_ns=list(mean_Z, var_Z, n_Z)} : summary statistics for non-selected sample
}
}
\description{
\code{ErrorCheck} examinate if the input data for calculating outcome mean estimates under proxy pattern-mixture model is correct.
}
\details{
\describe{
\item{\code{data_YZA=list(Y,Z)}(Microdata)}{
\itemize{
\item \code{Y} : \code{n*1} matrix of outcome variable \code{Y}
\item \code{Z} : \code{n*nZparams} matrix of auxiliary variables \code{Z}
}
}
\item{\code{data_YZA=list(mean_YZ, var_YZ, n_YZ)}(Summary Statistics)}{
\itemize{
\item \code{mean_YZ} : \code{(1+nZparams)*1} vector of mean (in order) of outcome variable \code{Y} and auxiliary variables \code{Z}
\item \code{var_YZ} : \code{(1+nZparams+nAparams)*(1+nZparams+nAparams)} matrix of variance (in order) of outcome variable \code{Y} and auxiliary variables \code{Z}
\item \code{n_YZ} : number of selected sample
}
}

\item{\code{data_Z_ns=list(Z)}(Microdata)}{
\itemize{
\item \code{Z} : \code{(N-n)*nZparams} matrix of auxiliary variables \code{Z}
}
}
\item{\code{data_Z_ns=list(mean_Z, var_Z, n_Z)}(Summary Statistics)}{
\itemize{
\item \code{mean_Z} : \code{(nZparams)*1} vector of mean (in order) of auxiliary variables \code{Z}
\item \code{var_Z} : \code{(nZparams)*(nZparams)} matrix of variance (in order) of auxiliary variables \code{Z}
\item \code{n_Z} : number of non-selected sample
}
}

}
}
