% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff_linear.R
\name{mub_coeff_linear_mle}
\alias{mub_coeff_linear_mle}
\title{Function to calculate MUB for non-binary outcomes}
\usage{
mub_coeff_linear_mle(data_YZA_s, data_ZA_ns, nZvars, phi = c(0, 0.5, 1))
}
\arguments{
\item{data_YZA_s}{list for selected sample (Microdata or Summary statistics). See details below for more details.}

\item{data_ZA_ns}{list for non-selected sample (Microdata or Summary statistics). See details below for more details.}

\item{nZvars}{dimension of predictor variables \code{Z}}

\item{phi}{scalar or vector of phi values at which the (S)MUB is calculated, should be in \verb{[0,1]} but values from \verb{[-Inf,1]} are allowed}
}
\value{
A dataframe with dimension of \code{ndraws*nZvars} with each row the values of MUB for each Z variable with the given phi (Column 1).
}
\description{
\code{mub_coeff_linear_mle} calculate maximum likelihood estimates of MUB for non-binary outcomes
}
\details{
\describe{
\item{\code{data_YZA_s=list(Y,Z,A)}(Microdata)}{
\itemize{
\item \code{Y} : \code{n*1} matrix of outcome variable \code{Y}
\item \code{Z} : \code{n*nZvars} matrix of predictor variables \code{Z}
\item \code{A} : \code{n*nAvars}matrix of auxiliary variables \code{A}
}
}
\item{\code{data_YZA_s=list(mean_YZA, var_YZA, n_YZA)}(Summary Statistics)}{
\itemize{
\item \code{mean_YZA} : \code{(1+nZvars+nAvars)*1} vector of mean (in order) of outcome variable \code{Y}, predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{var_YZA} : \code{(1+nZvars+nAvars)*(1+nZvars+nAvars)} matrix of variance (in order) of outcome variable \code{Y}, predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{n_YZA} : number of selected sample
}
}

\item{\code{data_ZA_ns=list(Z,A)}(Microdata)}{
\itemize{
\item \code{Z} : \code{(N-n)*nZvars} matrix of predictor variables \code{Z}
\item \code{A} : \code{(N-n)*nAvars}matrix of auxiliary variables \code{A}
}
}
\item{\code{data_ZA_ns=list(mean_ZA, var_ZA, n_ZA)}(Summary Statistics)}{
\itemize{
\item \code{mean_ZA} : \code{(nZvars+nAvars)*1} vector of mean (in order) of predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{var_ZA} : \code{(nZvars+nAvars)*(nZvars+nAvars)} matrix of variance (in order) of predictor variables \code{Z} and auxiliary variables \code{A}
\item \code{n_ZA} : number of non-selected sample
}
}

}
}
\examples{
require(mvtnorm)
set.seed(123)
# num of total sample
N <- 10000
nZvars <- 2
mu <- c(10, 0, 0, 0)
rhoy1 <- 0.4
rhoy2 <- 0.6
rho1a <- 0.2
sigmaya <- 0.3
Sigma <- matrix(c(
  4, 2 * rhoy1, 2 * rhoy2, sigmaya,
  2 * rhoy1, 1, 0, rho1a,
  2 * rhoy2, 0, 1, 0,
  sigmaya, rho1a, 0, 1
), 4, 4)
raw_data <- rmvnorm(N, mu, Sigma)
colnames(raw_data) <- c("Y", "Z1", "Z2", "A")
Y <- raw_data[, 1]
Z <- raw_data[, 2:3]
A <- raw_data[, 4]
prob <- plogis(-4 + Z \%*\% c(log(1.1), log(1.1)) + log(1.1) * Y + log(2) * A)
S <- rbinom(prob, 1, prob)
index_s <- S == 1
index_ns <- S == 0
# generate the list of microdata for selected (data_YZA_s) and non-selected sample (data_ZA_ns)
data_YZA_s <- list(Y = Y[index_s], Z = Z[index_s, ], A = A[index_s])
data_ZA_ns <- list(Z = Z[index_ns, ], A = A[index_ns])

# draw mle result from ppmm model
result <- mub_coeff_linear_mle(data_YZA_s, data_ZA_ns, nZvars)

# Use summary statistics for both selected and non-selected sample ------------------------
# generate the list of summary statistics for selected sample (data_YZA_sumry)
data_YZA_sumry <- list(
  mean_YZA = colMeans(raw_data[index_s, ]),
  var_YZA = var(raw_data[index_s, ]),
  n_YZA = sum(index_s)
)
# generate the list of summary statistics for non-selected sample (data_ZA_sumry)
data_ZA_sumry <- list(
  mean_ZA = colMeans(raw_data[index_ns, -1]),
  var_ZA = var(raw_data[index_ns, -1]),
  n_ZA = sum(index_ns)
)
result_sumry <- mub_coeff_linear_mle(data_YZA_sumry, data_ZA_sumry, nZvars)

}
\references{
West, B. T., Little, R. J., Andridge, R. R., Boonstra, P. S., Ware, E. B., Pandit, A., & Alvarado-Leiton, F. (2021). Assessing selection bias in regression coefficients estimated from nonprobability samples with applications to genetics and demographic surveys. \emph{The annals of applied statistics}, 15(3), 1556–1581.
}
