% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_sversion.R
\name{prop_bayes}
\alias{prop_bayes}
\title{Calculate Bayes estimates for binary outcomes}
\usage{
prop_bayes(data_YZ_s, data_Z_ns, phi = 0, phi_character = NULL, ndraws = 1000)
}
\arguments{
\item{data_YZ_s}{list for selected sample (Microdata). See details below for more details.}

\item{data_Z_ns}{list for non-selected sample (Microdata or Summary statistics). See details below for more details.}

\item{phi}{value of phi, fixed to a value between 0 and 1 (defaults to 0)}

\item{phi_character}{draw phi from specific distribution (i.e."\code{runif(1)}", defaults to \code{NULL})
\itemize{
\item If \code{phi_character=NULL}, fix at value set to \code{phi};
\item If \verb{phi_character=} specific distribution (i.e. \code{runif(1)} or \code{rbeta(1,1,1)}), then the value of \code{phi} input to the function is ignored
}}

\item{ndraws}{number of draws}
}
\value{
A \code{ndraws*17} matrix of Bayes estimates with each row containing:
\itemize{
\item \code{phi} : drawn phi in this draw
\item \code{pi}: drawn proportion of selected sample
\item \code{muX_s} : drawn mean of the proxy for selected sample
\item \code{muU_s} : drawn mean of the underlying latent variable for selected sample
\item \code{muY_s} : drawn mean of outcome variable for selected sample
\item \code{muX_ns} : drawn mean of the proxy for non-selected sample
\item \code{muU_ns} : drawn mean of the underlying latent variable for non-selected sample
\item \code{muY_ns} : drawn mean of outcome variable for non-selected sample
\item \code{sigmaXX_s} : drawn variance of the proxy for selected sample
\item \code{sigmaXU_s} : drawn covariance between the proxy and the underlying latent variable for selected sample
\item \code{sigmaUU_s} : drawn variance of the underlying latent variable for selected sample
\item \code{sigmaXX_ns} : drawn variance of the proxy for non-selected sample
\item \code{sigmaXU_ns} : drawn covariance between the proxy and the underlying latent variable for non-selected sample
\item \code{sigmaUU_ns} : drawn variance of the underlying latent variable for non-selected sample
\item \code{rho_s} : drawn biserial correlation for selected sample
\item \code{muY} : drawn population mean of outcome variable
\item \code{msb} : drawn MSB
}
}
\description{
Calculate Bayes estimates for binary outcomes
}
\details{
\describe{
\item{\code{data_YZ_s=list(Y,Z)}(Microdata)}{
\itemize{
\item \code{Y} : \code{n*1} matrix of outcome variable \code{Y}
\item \code{Z} : \code{n*nZparams} matrix of auxiliary variables \code{Z}
}
}
\item{\code{data_Z_ns=list(Z)}(Microdata)}{
\itemize{
\item \code{Z} : \code{(N-n)*nZparams} matrix of auxiliary variables \code{Z}
}
}
\item{\code{data_Z_ns=list(mean_Z, var_Z, n_Z)}(Summary Statistics)}{
\itemize{
\item \code{mean_Z} : \code{nZparams*1} vector of mean of auxiliary variables \code{Z}
\item \code{var_Z} : \code{nZparams*nZparams} matrix of variance of auxiliary variables \code{Z}
\item \code{n_Z} : number of non-selected sample
}
}
}
}
\examples{
set.seed(123)
# num of total sample
N <- 10000
# generate auxiliary variables Z
Z <- rnorm(N, 0, 1)
# define the biserial correlation rho between outcome variable Y and the proxy X
rho <- 0.5
# define the mean of the outcome variable Y
mu_y <- 0.3
# define a0, a1 such that E[Y] = mu_y
a1 <- rho / sqrt(1 - rho^2)
a0 <- qnorm(mu_y) * sqrt(1 + a1^2)
# generate the latent variable U
U <- rnorm(Z, a0 + a1 * Z, 1)
# generate the observed variable Y
Y <- ifelse(U > 0, 1, 0)
# generate the sample selection indicator S with 0.05 of the population selected
prob <- plogis(-3 + 0.3 * Z + 0.1 * U)
S <- rbinom(prob, 1, prob)
index_s <- S == 1
index_ns <- S == 0
# generate the list of microdata for selected (data_YZ_s) and non-selected sample (data_Z_ns)
data_YZ_s <- list(Y = Y[index_s], Z = Z[index_s])
data_Z_ns <- list(Z = Z[index_ns])

# draw bayesian result from ppmm model
result <- prop_bayes(data_YZ_s, data_Z_ns, phi_character = "runif(1)")

# Use summary statistics for non-selected sample ------------------------------------------
# generate the list of summary statistics for non-selected sample (data_Z_sumry)
data_Z_sumry <- list(
  mean_Z = mean(Z[index_ns]),
  var_Z = var(Z[index_ns]),
  n_Z = length(Z[index_ns])
)
result_sumry <- prop_bayes(data_YZ_s, data_Z_sumry, phi_character = "runif(1)")

}
\references{
Andridge, R. R., West, B. T., Little, R. J. A., Boonstra, P. S., & Alvarado-Leiton, F. (2019). Indices of non-ignorable selection bias for proportions estimated from non-probability samples. \emph{Journal of the Royal Statistical Society. Series C, Applied statistics, 68(5)}, 1465–1483.
}
\seealso{
\code{\link[=means_bayes]{means_bayes()}}

\code{\link[=means_mle]{means_mle()}}

\code{\link[=prop_mle]{prop_mle()}}
}
