% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_mi_sversion.R
\name{prop_mi}
\alias{prop_mi}
\title{Function for Proxy Pattern-Mixture Analysis (Binary Outcomes)}
\usage{
prop_mi(
  data_YZ_s,
  data_Z_ns,
  phi = 0,
  phi_character = NULL,
  D = 10,
  burnin = 500,
  thin = 100
)
}
\arguments{
\item{data_YZ_s}{list of Y, Z for selected sample (Microdata only)}

\item{data_Z_ns}{list of Z for non-selected sample (Microdata only)}

\item{phi}{value of phi, fixed to a value between 0 and 1 (defaults to 0)}

\item{phi_character}{draw phi from specific distribution (i.e."\code{runif(1)}", defaults to \code{NULL})
\itemize{
\item If \code{phi_character=NULL}, fix at value set to \code{phi};
\item If \verb{phi_character=} specific distribution (i.e. \code{runif(1)} or \code{rbeta(1,1,1)}), then the value of \code{phi} input to the function is ignored
}}

\item{D}{number of multiply imputed data sets (defaults to 10)}

\item{burnin}{number of burn-in replicates (defaults to 500)}

\item{thin}{number of replicates between imputations (defaults to 100)}
}
\value{
A \code{N*D} matrix of multiply imputed binary Y vectors, with each column an imputed outcome data set.
}
\description{
\code{prop_mi} calculate multiply imputed outcome for binary outcomes.
}
\details{
\describe{
\item{\code{data_YZ_s=list(Y,Z)}(Microdata)}{
\itemize{
\item \code{Y} : \code{n*1} matrix of outcome variable \code{Y}
\item \code{Z} : \code{n*nZparams} matrix of auxiliary variables \code{Z}
}
}
\item{\code{data_Z_ns=list(Z)}(Microdata)}{
\itemize{
\item \code{Z} : \code{(N-n)*nZparams} matrix of auxiliary variables \code{Z}
}
}
}
}
\examples{
set.seed(123)
# num of total sample
N <- 10000
# generate auxiliary variables Z
Z <- rnorm(N, 0, 1)
# define the biserial correlation rho between outcome variable Y and the proxy X
rho <- 0.5
# define the mean of the outcome variable Y
mu_y <- 0.3
# define a0, a1 such that E[Y] = mu_y
a1 <- rho / sqrt(1 - rho^2)
a0 <- qnorm(mu_y) * sqrt(1 + a1^2)
# generate the latent variable U
U <- rnorm(Z, a0 + a1 * Z, 1)
# generate the observed variable Y
Y <- ifelse(U > 0, 1, 0)
# generate the sample selection indicator S with 0.05 of the population selected
prob <- plogis(-3 + 0.3 * Z + 0.1 * U)
S <- rbinom(prob, 1, prob)
index_s <- S == 1
index_ns <- S == 0
# generate the list of microdata for selected (data_YZ_s) and non-selected sample (data_Z_ns)
data_YZ_s <- list(Y = Y[index_s], Z = Z[index_s])
data_Z_ns <- list(Z = Z[index_ns])

# impute the missing outcome data
imp <- prop_mi(data_YZ_s, data_Z_ns, phi = 0, D = 10, burnin = 500, thin = 100)
}
\seealso{
\code{\link[=cont_mi]{cont_mi()}}
}
