% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_sversion.R
\name{prop_mle}
\alias{prop_mle}
\title{Calculate MUBP estimates for binary outcomes with maximum likelihood estimation}
\usage{
prop_mle(data_YZ_s, data_Z_ns, phi = c(0, 0.5, 1), sfrac = NULL)
}
\arguments{
\item{data_YZ_s}{list for selected sample (Microdata). See details below for more details.}

\item{data_Z_ns}{list for non-selected sample (Microdata or Summary statistics). See details below for more details.}

\item{phi}{value of phi with a value between 0 and 1}

\item{sfrac}{if sfrac is not NULL, the proportion of selected sample is fixed to \code{sfrac}, else it is estimated from the data (the proportion of selected sample).}
}
\value{
A \code{length(phi)*16} matrix of Bayes estimates with each row containing:
\itemize{
\item \code{phi} : corresponding phi
\item \code{pi}: estimated proportion of selected sample (or defined by user)
\item \code{muX_s} : ML estimate of the mean of the proxy for selected sample
\item \code{muU_s} : ML estimate of the mean of the underlying latent variable for selected sample
\item \code{muY_s} : ML estimate of the mean of outcome variable for selected sample
\item \code{muX_ns} : ML estimate of the mean of the proxy for non-selected sample
\item \code{muU_ns} : ML estimate of the mean of the underlying latent variable for non-selected sample
\item \code{muY_ns} : ML estimate of the mean of outcome variable for non-selected sample
\item \code{sigmaXX_s} : ML estimate of the variance of the proxy for selected sample
\item \code{sigmaXU_s} : ML estimate of the covariance between the proxy and the underlying latent variable for selected sample
\item \code{sigmaUU_s} : ML estimate of the variance of the underlying latent variable for selected sample
\item \code{sigmaXX_ns} : ML estimate of the variance of the proxy for non-selected sample
\item \code{sigmaUU_ns} : ML estimate of the variance of the underlying latent variable for non-selected sample
\item \code{rho_s} : ML estimate of the biserial correlation for selected sample
\item \code{muY} : ML estimate of the population mean of outcome variable
\item \code{msb} : ML estimate of the MSB
}
}
\description{
Calculate MUBP estimates for binary outcomes with maximum likelihood estimation
}
\details{
\describe{
\item{\code{data_YZ_s=list(Y,Z)}(Microdata)}{
\itemize{
\item \code{Y} : \code{n*1} matrix of outcome variable \code{Y}
\item \code{Z} : \code{n*nZparams} matrix of auxiliary variables \code{Z}
}
}
\item{\code{data_Z_ns=list(Z)}(Microdata)}{
\itemize{
\item \code{Z} : \code{(N-n)*nZparams} matrix of auxiliary variables \code{Z}
}
}
\item{\code{data_Z_ns=list(mean_Z, var_Z, n_Z)}(Summary Statistics)}{
\itemize{
\item \code{mean_Z} : \code{nZparams*1} vector of mean of auxiliary variables \code{Z}
\item \code{var_Z} : \code{nZparams*nZparams} matrix of variance of auxiliary variables \code{Z}
\item \code{n_Z} : number of non-selected sample
}
}
}
}
\examples{
set.seed(123)
# num of total sample
N <- 10000
# generate auxiliary variables Z
Z <- rnorm(N, 0, 1)
# define the biserial correlation rho between outcome variable Y and the proxy X
rho <- 0.5
# define the mean of the outcome variable Y
mu_y <- 0.3
# define a0, a1 such that E[Y] = mu_y
a1 <- rho / sqrt(1 - rho^2)
a0 <- qnorm(mu_y) * sqrt(1 + a1^2)
# generate the latent variable U
U <- rnorm(Z, a0 + a1 * Z, 1)
# generate the observed variable Y
Y <- ifelse(U > 0, 1, 0)
# generate the sample selection indicator S with 0.05 of the population selected
prob <- plogis(-3 + 0.3 * Z + 0.1 * U)
S <- rbinom(prob, 1, prob)
index_s <- S == 1
index_ns <- S == 0
# generate the list of microdata for selected (data_YZ_s) and non-selected sample (data_Z_ns)
data_YZ_s <- list(Y = Y[index_s], Z = Z[index_s])
data_Z_ns <- list(Z = Z[index_ns])

## calculate MUB result from ppmm model with maximum likelihood estimation
result <- prop_mle(data_YZ_s, data_Z_ns)

# Use summary statistics for non-selected sample ------------------------------------------
# generate the list of summary statistics for non-selected sample (data_Z_sumry)
data_Z_sumry <- list(
  mean_Z = mean(Z[index_ns]),
  var_Z = var(Z[index_ns]),
  n_Z = length(Z[index_ns])
)
result_sumry <- prop_mle(data_YZ_s, data_Z_sumry)

}
\references{
Andridge, R. R., West, B. T., Little, R. J. A., Boonstra, P. S., & Alvarado-Leiton, F. (2019). Indices of non-ignorable selection bias for proportions estimated from non-probability samples. \emph{Journal of the Royal Statistical Society. Series C, Applied statistics, 68(5)}, 1465–1483.
}
\seealso{
\code{\link[=means_bayes]{means_bayes()}}

\code{\link[=means_mle]{means_mle()}}

\code{\link[=prop_bayes]{prop_bayes()}}
}
